fixing integer overflow errors; formatting code stuff too

+++ src/HyDe.hpp
@@ -30,20 +30,20 @@ private:
 
   /* Functions for calculating stuff.*/
   inline int _convert(char str);                                                 /* Convert DNA bases form characters to ints.      */
-  int _resolveAmbiguity(const int& out, const int& p1,                           /* Randomly resolves ambiguity codes.              */
-                        const int& hyb, const int& p2,
-                        double cp[16][16]);
-  double _calcGH(const double cp[16][16], const int& nObs,
-                 const int& avgObs, const int& p1,
+  double _resolveAmbiguity(const int& out, const int& p1,                        /* Randomly resolves ambiguity codes.              */
+                           const int& hyb, const int& p2,
+                           double cp[16][16]);
+  double _calcGH(const double cp[16][16], const double& nObs,
+                 const double& avgObs, const int& p1,
                  const int& hyb, const int& p2);                                 /* Calculates GH test statistic.                   */
   double _calcPvalue(const double& myZ);                                         /* Calculate p-value with bonferroni correction.   */
   double _calcPvalueTwo(const double& myZ);
   inline long int _nCk(long int& n, long int& k);                                /* Calculates binomial coefficient: n choose k.    */
   double _bonferroniCorrect();                                                   /* Bonferroni correction based on number of taxa.  */
-  int _getCountMatrix(const int& p1,                                             /* Populates count matrix from given triplet.      */
-                      const int& hyb,
-                      const int& p2,
-                      double cp[16][16]);
+  double _getCountMatrix(const int& p1,                                          /* Populates count matrix from given triplet.      */
+                         const int& hyb,
+                         const int& p2,
+                         double cp[16][16]);
 
   /* private member variables */
   std::string _infile = "none", _mapfile = "none", _prefix = "hyde";
